将motor_set_duty函数的参数类型从int8_t更改为int16_t，以支持更大的占空比范围

+++ main/motor.cpp
@@ -40,7 +40,7 @@ void motor_control_init()
     ESP_ERROR_CHECK(gpio_config(&motor_gpio_cfg));
 }
 
-void motor_set_duty(int8_t duty)
+void motor_set_duty(int16_t duty)
 {
     if(duty >= 0)
     {
